requirements.txt
# Core MLOps and ClearML
clearml>=1.13.0
clearml-agent>=1.7.0

# Data Science and ML
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Database and ETL
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.8.0

# Deep Learning (optional)
# torch>=1.12.0
# tensorflow>=2.8.0

# Utilities
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Development and Testing
pytest>=7.0.0
jupyter>=1.0.0
ipython>=8.0.0

# Model deployment (optional)
# mlflow>=1.28.0
# fastapi>=0.80.0
# uvicorn>=0.18.0

kagglehub==0.3.12
